It WORKS!
JFC it took me an entire day to figure out why my little dudes weren't moving anymore.  turns out just adding this: [SerializeField] private NetworkPrefabRef unitActionSystemPrefab; and then turning my UnitActionSystem into a prefab to ref it and then making sure that players spwned with input authority was the answer. That was painful and awful, but it works and I can move on to something ridiculous that is stupid easy for most but will drive me crazy.