MOVEMENT FINALLT!
It's ugly AF but I can move units with the Host and with the remote client.  there are many problems though, but I had to break things to try and get things working lol. the units move on their own towards 0,0,0, unless you're holding down the right mouse button. it's kind of funny.