Added Animation For Player walking
cleaned up a lot of the network logic to ensure IsSeelcted is sync'd across the network and can be used to ensure a unit is selected before it's moved. Removed a lot of redundancy and completely removed UnitActionSystem.cs (commented out) and moved that logic into the Unit.cs for simplicity and to fix some small issues. Units now will move to destination if deselected mid movement.