Health System added
Added health system and very basic death logic. game has a few bugs. when things are desapawned it can cause a null if we don't remove the event subscriptions and lists then the game will bug out. it works but if things get out of sequence then it will bug throw a null. Have override for Despawn in grid position, but TeamID's are still an issue it looks like