Fixed Ragdoll and Death Behavior
Removed the state authority check to allow clients to spawn their own ragdoll prefab then RPC sync to the other players. still wonky physics need to be addressed for the force applied to ragdoll on death.. Added a bunch of extension methods written by git-amend. Haven't used them in the game so far, but I like a lot of them so I'll try to learn them and use them to clean up some clutter