Death Ragdoll Added
Added the ragdoll logic to spawn in as a 1 for 1 of the killed unit. The coroutine of removing the unit is causing a brief visual weirdness where it looks like 2 units are spawned in at once. I'll have to tweak this timing down. but without making it completely networked there may just have to be some visual I'll have to live with.