Added Health Bar
Added UI health bar and a network behavior script to visually sync it on the network. Things are getting too heave on Unit.cs so I moved the damage UI logic to a new class. Probably can move the death logic as well, but just gonna leave it for now.