Added Fractured Crate
Modified my destructible crate model in Blender with RBDLab to simulate an explosion. it's not super pretty but it works. Linked that effect on the network and now my crate bursts into pieces when the grenade hits it. 865 or so fragments though, so I think I need to pool the effect or it could slow the game down.