Updated Grid Visuals
Host and Client have visual feedback for grids containing enemy units for a given action. It's not perfect because it kind of bugs out when the units move. it fixes itself when you move your own unit though, so I think it's a matter of a refresh/update call somewhere