Fixed Knife Attack Feedback
Turns out all I had to do was redrag in the gameObject. I'm glad it took me an entire day to figure out it wasn't my logic that was flawed, just Fusion... good times. Cleaned up the timing of the feedback to play the sound early in the sequence and then shake the camera on impact. Though it's not really on impact it's just a timing thing.